requirements.txt
asgiref==3.5.2
azure-common==1.1.27
azure-core==1.24.1
azure-functions==1.11.2
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
azure-storage-blob==12.9.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
isodate==0.6.1
msal==1.18.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.0
Pillow==9.1.1
portalocker==2.4.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
pywin32==304
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
